pallet-tracker/requirements.txt
streamlit
google-cloud-vision
gspread
oauth2client
pandas
